refactor almanac
This method lacks the randomness.  Instead, it just takes the current
day in Fenestra as a starting number, then ticks up, then returns to the
start.

As a result, each mundane day, three encounters are skipped, which means
the next compile will probably look similar, but the days won't match
1:1 to a previous compile's version of what happens on which day.